+++ packages/wonder-blocks-pill/src/index.ts
@@ -1,2 +1,3 @@
-// Just a temporary export to have some code.
-export const NOTHING = "NOTHING";
+import Pill from "./components/pill";
+
+export default Pill;