+++ packages/wonder-blocks-pill/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @khanacademy/wonder-blocks-pill
 
+## 1.0.0
+
+### Major Changes
+
+-   3455afae: Create Pill component
+
+## 0.0.3
+
+### Patch Changes
+
+-   Updated dependencies [8dc4a5a3]
+-   Updated dependencies [1920feb8]
+    -   @khanacademy/wonder-blocks-clickable@4.0.2
+    -   @khanacademy/wonder-blocks-core@6.0.1
+    -   @khanacademy/wonder-blocks-link@5.0.2
+    -   @khanacademy/wonder-blocks-typography@2.1.4
+
 ## 0.0.2
 
 ### Patch Changes